[PE-6800] Add frosted component for button action bar
